requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
gunicorn==22.0.0
pytest==8.3.2
